First commit of this module
It parses a mediainfo stdout and wraps it into a mediainfo object

It only consists into a spawnSync call, a async method will be added later on.

Also, no crash handling is done, neither unit tests.

This is just a beta version for now, but feel free to test it and provide any feedback :)